........
  r706772 | josh | 2008-10-21 14:25:50 -0700 (Tue, 21 Oct 2008) | 1 line
  
  Fix for bug 46053 - fixed evaluation cache dependency analysis when changing blank cells
........
  r707450 | josh | 2008-10-23 12:08:42 -0700 (Thu, 23 Oct 2008) | 1 line
  
  Fix for bug 45778 - made ObjRecord read sub-record ftLbsData properly
........
  r707481 | josh | 2008-10-23 14:42:05 -0700 (Thu, 23 Oct 2008) | 1 line
  
  fixed re-serialization of tAttrChoose
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@707942 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/ArrayPtg.java
@@ -50,7 +50,7 @@ public final class ArrayPtg extends Ptg {
 	private final byte[] field_1_reserved;
 	
 	// data from these fields comes after the Ptg data of all tokens in current formula
-	private short  token_1_columns;
+	private int  token_1_columns;
 	private short token_2_rows;
 	private Object[] token_3_arrayValues;
 
@@ -109,7 +109,7 @@ public final class ArrayPtg extends Ptg {
 	 * See page 304-305 of Excel97-2007BinaryFileFormat(xls)Specification.pdf
 	 */
 	public void readTokenValues(RecordInputStream in) {
-		short nColumns = in.readUByte();
+		int nColumns = in.readUByte();
 		short nRows = in.readShort();
 		//The token_1_columns and token_2_rows do not follow the documentation.
 		//The number of physical rows and columns is actually +1 of these values.
@@ -172,7 +172,7 @@ public final class ArrayPtg extends Ptg {
 	}
 
 	public short getColumnCount() {
-		return token_1_columns;
+		return (short)token_1_columns;
 	}
 
 	/** This size includes the size of the array Ptg plus the Array Ptg Token value size*/
